+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - a new view that shows all in-progress projects at `/projects/all/in-progress`
+- a new view that shows all in-progress projects assigned to Regional casework
+  services at `/projects/regional-casework-services/in-progress`
 
 ## [Release 16][release-16]